PYIC-3541 Rename ci-scoring resources
