+++ deploy/criReturnStepFunction.asl.json
@@ -37,9 +37,9 @@
       "ResultPath": "$.lambdaResult",
       "Next": "RouteJourneyResponse"
     },
-    "CiScoring": {
+    "CheckCiScore": {
       "Type": "Task",
-      "Resource": "${CiScoringFunctionArn}",
+      "Resource": "${CheckCiScoreFunctionArn}",
       "Parameters": {
         "ipvSessionId.$": "$.ipvSessionId",
         "ipAddress.$": "$.ipAddress",
@@ -88,7 +88,7 @@
         {
           "Variable": "$.lambdaResult.journey",
           "StringEquals": "/journey/ci-scoring",
-          "Next": "CiScoring"
+          "Next": "CheckCiScore"
         },
         {
           "Variable": "$.lambdaResult.journey",
