+++ lambdas/check-ci-score/src/main/resources/log4j2.json
@@ -1,0 +1,32 @@
+{
+  "Configuration": {
+    "status": "warn",
+    "appenders": {
+      "Console": {
+        "name": "JsonAppender",
+        "target": "SYSTEM_OUT",
+        "JsonTemplateLayout": {
+          "eventTemplateUri": "classpath:IpvLambdaJsonLayout.json"
+        }
+      }
+    },
+    "Loggers": {
+      "logger": [
+        {
+          "name": "JsonLogger",
+          "level": "info",
+          "additivity": false,
+          "AppenderRef": {
+            "ref": "JsonAppender"
+          }
+        }
+      ],
+      "Root": {
+        "level": "info",
+        "AppenderRef": {
+          "ref": "JsonAppender"
+        }
+      }
+    }
+  }
+}
