--- lambdas/ci-scoring/build.gradle
@@ -1,55 +1,0 @@
-plugins {
-	id "java"
-	id "idea"
-	id "jacoco"
-	id 'io.freefair.aspectj.post-compile-weaving' version '8.0.1'
-}
-
-repositories {
-	mavenCentral()
-}
-
-dependencies {
-	implementation "com.amazonaws:aws-java-sdk-sqs:$rootProject.ext.dependencyVersions.awsJavaSdkSqs",
-			"com.amazonaws:aws-lambda-java-core:$rootProject.ext.dependencyVersions.awsLambdaJavaCore",
-			"com.amazonaws:aws-lambda-java-events:$rootProject.ext.dependencyVersions.awsLambdaJavaEvents",
-			"com.nimbusds:oauth2-oidc-sdk:$rootProject.ext.dependencyVersions.nimbusdsOauth2OidcSdk",
-			project(":libs:common-services"),
-			project(":libs:cimit-service"),
-			project(":libs:gpg45-evaluator")
-
-	implementation project(path: ':libs:journey-uris')
-
-	aspect "software.amazon.lambda:powertools-logging:$rootProject.ext.dependencyVersions.powertoolsLogging",
-			"software.amazon.lambda:powertools-tracing:$rootProject.ext.dependencyVersions.powertoolsTracing"
-	aspect('org.aspectj:aspectjrt') {
-		version {
-			strictly '1.9.8'
-		}
-	}
-
-	testImplementation "com.google.code.gson:gson:$rootProject.ext.dependencyVersions.gson",
-			"org.junit.jupiter:junit-jupiter:5.10.0",
-			"org.mockito:mockito-junit-jupiter:5.4.0",
-			project(path: ':libs:common-services', configuration: 'tests')
-	testRuntimeOnly("org.junit.platform:junit-platform-launcher")
-}
-
-java {
-	sourceCompatibility = JavaVersion.VERSION_17
-	targetCompatibility = JavaVersion.VERSION_17
-}
-
-test {
-	// Configures environment variable to avoid initialization of AWS X-Ray segments for each tests
-	environment "LAMBDA_TASK_ROOT", "handler"
-	useJUnitPlatform ()
-	finalizedBy jacocoTestReport
-}
-
-jacocoTestReport {
-	dependsOn test
-	reports {
-		xml.required.set(true)
-	}
-}
