+++ libs/common-services/src/main/java/uk/gov/di/ipv/core/library/config/ConfigurationVariable.java
@@ -14,7 +14,7 @@ public enum ConfigurationVariable {
     PUBLIC_KEY_MATERIAL_FOR_CORE_TO_VERIFY("clients/%s/publicKeyMaterialForCoreToVerify"),
     CLIENT_VALID_REDIRECT_URLS("clients/%s/validRedirectUrls"),
     CI_SCORING_CONFIG("self/ci-scoring-config"),
-    CI_SCORING_THRESHOLD("self/ciScoringThreshold"),
+    CI_SCORING_THRESHOLD("self/checkCiScoreThreshold"),
     VC_TTL("self/vcTtl"),
     CREDENTIAL_ISSUERS("credentialIssuers"),
     FEATURE_FLAGS("featureFlags/%s"),
